+++ surveys/apps.py
@@ -1,4 +1,4 @@
-rom django.apps import AppConfig
+from django.apps import AppConfig
 
 
 class SurveysConfig(AppConfig):
